requirements.txt
argcomplete==3.4.0
asgiref==3.8.1
astroid==3.2.4
boto3==1.34.146
botocore==1.34.146
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
defusedxml==0.7.1
dill==0.3.8
Django==5.0.7
django-better-admin-arrayfield==1.4.2
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-json-widget==2.0.1
django-rest-swagger==2.2.0
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
jira==3.8.0
jmespath==1.0.1
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==2.0.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pandas==1.5.2
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pylint==3.2.6
pylint-report==2.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.10.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
python-decouple==3.8